Move common cli helpers to own folder
