+++ internals/scripts/helpers/checkmark.js
@@ -1,0 +1,12 @@
+/* eslint-disable */
+var chalk = require('chalk');
+
+/**
+ * Adds mark check symbol
+ */
+function addCheckMark(callback) {
+  process.stdout.write(chalk.green(' ✓'));
+  callback();
+}
+
+module.exports = addCheckMark;
