+++ internals/scripts/helpers/progress.js
@@ -1,0 +1,24 @@
+/* eslint-disable */
+
+/**
+ * Adds an animated progress indicator
+ *
+ * @param  {string} message      The message to write next to the indicator
+ * @param  {number} amountOfDots The amount of dots you want to animate
+ */
+function animateProgress(message, amountOfDots) {
+  if (typeof amountOfDots !== 'number') {
+    amountOfDots = 3;
+  }
+
+  var i = 0;
+  return setInterval(function () {
+    process.stdout.clearLine();
+    process.stdout.cursorTo(0);
+    i = (i + 1) % (amountOfDots + 1);
+    var dots = new Array(i + 1).join(".");
+    process.stdout.write(message + dots);
+  }, 500);
+}
+
+module.exports = animateProgress;
